requirements.txt
supabase==1.0.4
requests==2.31.0
python-dotenv==1.0.0
schedule==1.2.0
psycopg2-binary==2.9.9
